+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - [#2424] DefineEditText handling of letterSpacing, font size on incorrect values
 - [#2391] Double not operator in ternar operator expression
 - [#2436] PDF export clipping - missing colors / text
+- [#2437] AS1 P-code - do not group pushes automatically
+- [#2437] AS1 Direct editation - group pushes on FP5+
+- [#2437] AS1 - use Constant pool only on FP5+
 
 ## [22.0.2] - 2025-01-17
 ### Added
@@ -3699,6 +3702,7 @@ Major version of SWF to XML export changed to 2.
 [#2424]: https://www.free-decompiler.com/flash/issues/2424
 [#2391]: https://www.free-decompiler.com/flash/issues/2391
 [#2436]: https://www.free-decompiler.com/flash/issues/2436
+[#2437]: https://www.free-decompiler.com/flash/issues/2437
 [#2375]: https://www.free-decompiler.com/flash/issues/2375
 [#2374]: https://www.free-decompiler.com/flash/issues/2374
 [#2389]: https://www.free-decompiler.com/flash/issues/2389